[chains] Fixes: status polling timeout, unify example code.
